Better base instances
Fixed some searching parameters, currently the search converts the given comment into a list separated by whitespace and runs it against a name in the list. This means that things like "los angeles" in the name field would not catch because it would be ran against ["los", "angeles"]. I could fix this, however I do not think it is required as if someone types in "angeles" I am going to assume that they mean los angeles afb, thus I simply refined names with white spaces.