Added triggers for a bot to reply
Bot will only reply if a user is trying to rate a base or if the post contains !AFBbot / AFBbot / /u/AFBbot. This will prevent false positives as it is unlikely a post will contain "rate", an integer and a base name without someone trying to rate it (though it could be implemented here as well. Usage: !AFBbot Langley