Fixed bot replying to quoted text and decimals/negative numbers.
Bot will ignore quoted text and still try to reply to a comment as long as there is only one line of quoted text, otherwise it will ignore it.

Bot will only reply if a user is trying to rate a base or if the post contains bot name. This will prevent false positives as it is unlikely a post will contain "rate", an integer and a base name without someone trying to rate it (though it could be implemented here as well. Usage: !AFBbot Langley

Bot will now accept negative numbers, decimals, zero and any other number and in most formats (9/10, 9 out of 10). It will round to a number between 1-10.