Rebuilt filtering for ratings/quotes.
Posts are now filtered much better, bot will ignore quoted text. If there is only one line of quoted text the bot will still read the rest of the post, if there are multiple lines it will not. Rating filtering is rebuilt, it's a bit of spaghetti code but seems to handle well. Bot will only look at numbers that are after the word "rate", and will pick the first one. Bot also filters out extra periods or garbage in floats.

One issue to fix is that if the rating number is before the word "rate" it throws index out of range exception that is caught, this can be fixed in the future but is not a big deal.